Split event-datetime block into subblocks
