+++ src/modules/blocks/event-datetime/subblocks/controls/index.js
@@ -1,0 +1,1 @@
+export { default } from './container';
