+++ src/modules/blocks/event-datetime/subblocks/index.js
@@ -1,0 +1,20 @@
+/**
+ * External Dependencies
+ */
+import { registerBlockType } from '@wordpress/blocks';
+
+/**
+ * Internal Dependencies
+ */
+import content from './content';
+import dashboard from './dashboard';
+
+const blocks = [
+	content,
+	dashboard,
+];
+
+blocks.forEach( block => {
+	const blockName = `tribe/${ block.id }`;
+	registerBlockType( blockName, block );
+} );
